fix(sqlsmith): split weight config to config.rs
